Feat/tests (#17)
# Summary

The primary driver of this work was to introduce the `pytest` framework
to this project. This was a small change. It exploded into a full
re-architecting as the bot was invoke on execution which means when the
tests ran the client was being invoked and waiting for `SIGTERM` to
close the client and allow the tests to finish.

Taking this opportunity, I choose to introduce a CLI component to make
development easier by avoiding need to spin up the bot and execute via
Discord. This also provide an opportunity to refactor the function
location to a separate file, making `main.py` to read.